requirements.txt
# Core Django dependencies
Django>=5.2.0
django-cors-headers>=4.0.0

# Django REST Framework for API
djangorestframework>=3.14.0

# Optional: for better API documentation
drf-spectacular>=0.27.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
